Fixed iterative algorithm C implementation.

+++ c/ackermann.c
@@ -2,34 +2,34 @@
 #include <stdlib.h>
 #include "ackermann.h"
 
-int a(int m, int n) {
+long a(int m, int n) {
   if (m == 0) 
-    return n + 1;
+    return (long) n + 1;
   else if (n == 0) 
     return a(m-1, 1);
   else
     return a(m-1, a(m, n-1));
 }
 
-int memoized_a(int m, int n) {
+long memoized_a(int m, int n) {
   return  -1;
 }
 
-int iterative_a(int i, int n) {
-  int *next = calloc(i+1, sizeof(int));
-  int *next_c = next;
-  int *goal = calloc(i+1, sizeof(int));
+long iterative_a(int i, int n) {
+  long *next = calloc((long) i+1, sizeof(long));
+  long *next_c = next;
+  long *goal = calloc((long) i+1, sizeof(long));
   int idx = 0;
   while (idx < i+1){
     goal[idx++] = 1;
   }
   goal[i] = -1;
-  int *goal_c = goal;
-  int value = -1;
+  long *goal_c = goal;
+  long value = -1;
   do  {
     value = next[0] + 1;
     bool transferring = true;
-    int current = i;
+    long current = i;
     while (transferring) {
       if (next[i-current] == goal[i-current]) {
         goal[i-current] = value;
@@ -40,7 +40,7 @@ int iterative_a(int i, int n) {
       current -= 1;
     }
   } 
-  while (next[i] != n + 1);
+  while (next[i] != (long) n + 1);
   free(next_c);
   free(goal_c);
   return value;
